Removed the duplicate server files they're the same!
Andoid had the same content just updated
